+++ epoxy-processor/src/main/java/com/airbnb/epoxy/processor/PackageConfigSettings.kt
@@ -14,14 +14,15 @@ class PackageConfigSettings private constructor(
     companion object {
 
         fun forDefaults() = PackageConfigSettings(
-                PackageEpoxyConfig.REQUIRE_HASHCODE_DEFAULT,
-                PackageEpoxyConfig.REQUIRE_ABSTRACT_MODELS_DEFAULT,
-                PackageEpoxyConfig.IMPLICITLY_ADD_AUTO_MODELS_DEFAULT
+            PackageEpoxyConfig.REQUIRE_HASHCODE_DEFAULT,
+            PackageEpoxyConfig.REQUIRE_ABSTRACT_MODELS_DEFAULT,
+            PackageEpoxyConfig.IMPLICITLY_ADD_AUTO_MODELS_DEFAULT
         )
 
         fun create(configAnnotation: PackageEpoxyConfig) = PackageConfigSettings(
-                configAnnotation.requireHashCode,
-                configAnnotation.requireAbstractModels,
-                configAnnotation.implicitlyAddAutoModels)
+            configAnnotation.requireHashCode,
+            configAnnotation.requireAbstractModels,
+            configAnnotation.implicitlyAddAutoModels
+        )
     }
 }