+++ epoxy-processor/src/test/java/com/airbnb/epoxy/PoetExtensionsTest.kt
@@ -4,13 +4,13 @@ import androidx.annotation.FloatRange
 import androidx.annotation.NonNull
 import com.airbnb.epoxy.EpoxyModelClass
 import com.squareup.kotlinpoet.asTypeName
-import javax.lang.model.element.Modifier
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertNotNull
 import org.junit.Assert.assertNull
 import org.junit.Assert.assertTrue
 import org.junit.Test
+import javax.lang.model.element.Modifier
 
 class PoetExtensionsTest {