+++ epoxy-processortest/src/test/resources/DataBindingConfig.java
@@ -1,0 +1,4 @@
+package com.airbnb.epoxy;
+
+@EpoxyDataBindingLayouts({R.layout.model_with_data_binding})
+interface EpoxyDataBindingConfig {}