+++ epoxy-sample/src/main/java/com/airbnb/epoxy/sample/package-info.java
@@ -1,6 +1,4 @@
-@EpoxyDataBindingLayouts(R.layout.button)
 @PackageModelViewConfig(rClass = R.class)
 package com.airbnb.epoxy.sample;
 
-import com.airbnb.epoxy.EpoxyDataBindingLayouts;
 import com.airbnb.epoxy.PackageModelViewConfig;